test(strip): merge validate_annotations tests into test_strip_artifacts
The validate_annotations module was folded into strip_artifacts in
c55f836. Now that both call sites are unified, consolidate the two
test files: append the 7 validation tests to test_strip_artifacts
and delete tests/test_validate_annotations.py.

All 21 merged tests reuse the existing _strip helper (dropping the
duplicate _validate helper and the aliased import). Suite still
passes at 457.
