+++ tests/test_strip_artifacts.py
@@ -162,3 +162,56 @@ def test_strip_artifacts_rewrites_dangling_cython_annotation():
     assert "from typing import Any" in result
     assert "bint" not in result
     assert "def f(x: Any) -> None" in result
+
+
+def test_defined_annotation_is_untouched():
+    result = _strip("class Foo: ...\ndef f(x: Foo) -> Foo: ...")
+
+    assert "def f(x: Foo) -> Foo" in result
+    assert "Any" not in result
+
+
+def test_dangling_annotation_is_replaced_with_any():
+    result = _strip("def f(x: StrippedCimport) -> None: ...")
+
+    assert "from typing import Any" in result
+    assert "def f(x: Any) -> None" in result
+
+
+def test_nested_generic_dangling_argument_is_replaced():
+    result = _strip(
+        "from typing import List\ndef f(x: List[DanglingName]) -> None: ..."
+    )
+
+    assert "from typing import Any" in result
+    assert "List[Any]" in result
+
+
+def test_multiple_dangling_names_are_replaced():
+    result = _strip("def f(x: MissingOne) -> MissingTwo: ...")
+
+    assert "from typing import Any" in result
+    assert "def f(x: Any) -> Any" in result
+
+
+def test_builtin_annotations_are_preserved():
+    result = _strip("def f(x: int, y: str, z: bool) -> bytes: ...")
+
+    assert "def f(x: int, y: str, z: bool) -> bytes" in result
+    assert "Any" not in result
+
+
+def test_existing_any_import_is_not_duplicated():
+    result = _strip("from typing import Any\ndef f(x: Missing) -> Any: ...")
+
+    assert result.count("from typing import Any") == 1
+    assert "def f(x: Any) -> Any" in result
+
+
+def test_dangling_decorator_argument_is_replaced_with_any():
+    result = _strip(
+        "some_decorator = None\n@some_decorator(UndefinedType)\ndef foo() -> None: ..."
+    )
+
+    assert "from typing import Any" in result
+    assert "@some_decorator(Any)" in result
